+++ src/react-native.ts
@@ -1,4 +1,4 @@
-import { EdgeOtherMethods } from 'edge-core-js/types'
+import { MemletConfig } from 'memlet'
 import { NativeModules } from 'react-native'
 
 const { EdgeCurrencyPluginsModule } = NativeModules
@@ -7,6 +7,20 @@ const { sourceUri } = EdgeCurrencyPluginsModule.getConstants()
 export const pluginUri = sourceUri
 export const debugUri = 'http://localhost:8081/edge-currency-plugins.js'
 
-export function makePluginIo(): EdgeOtherMethods {
-  return {}
+export interface EdgeCurrencyPluginIoOptions {
+  readonly memletConfig?: MemletConfig
+}
+
+export interface EdgeCurrencyPluginNativeIo {
+  readonly memletConfig?: MemletConfig
+}
+
+export function makePluginIo(
+  options: EdgeCurrencyPluginIoOptions = {}
+): EdgeCurrencyPluginNativeIo {
+  const { memletConfig } = options
+
+  return {
+    memletConfig
+  }
 }